Move useFills to frontity package
