+++ packages/hooks/index.ts
@@ -1,2 +1,1 @@
-export { default as useFills } from "./use-fills";
 export { default as useInView } from "./use-in-view";
